+++ Gw2Sharp/WebApi/V2/Clients/Commerce/CommerceExchangeCoinsClient.cs
@@ -1,5 +1,4 @@
 using System;
-using Gw2Sharp.WebApi.V2.Models;
 
 namespace Gw2Sharp.WebApi.V2.Clients
 {
@@ -7,7 +6,7 @@ namespace Gw2Sharp.WebApi.V2.Clients
     /// A client of the Guild Wars 2 API v2 commerce exchange coins endpoint.
     /// </summary>
     [EndpointPath("commerce/exchange/coins")]
-    public class CommerceExchangeCoinsClient : BaseEndpointBlobClient<CommerceExchangeCoins>, ICommerceExchangeCoinsClient
+    public class CommerceExchangeCoinsClient : BaseClient, ICommerceExchangeCoinsClient
     {
         /// <summary>
         /// Creates a new <see cref="CommerceExchangeCoinsClient"/> that is used for the API v2 commerce exchange coins endpoint.